feat: extract ScreenHead function

+++ shared/src/main/kotlin/com/example/pizzashift/shared/domain/model/OrderedPizzaIngredient.kt
@@ -2,5 +2,5 @@ package com.example.pizzashift.shared.domain.model
 
 data class OrderedPizzaIngredient(
     val name: IngredientType,
-    val cost: Int
+    val cost: Double
 )
